Updated file structure in src
