+++ src/rcpp_dev_parse_xml.cpp
@@ -13,8 +13,12 @@
 #include <algorithm>
 #include <iterator>
 
+
+
+
+
 // [[Rcpp::export]]
-Rcpp::List rcpp_parse_xml(Rcpp::NumericVector input) {
+Rcpp::List rcpp_dev_parse_xml(Rcpp::NumericVector input) {
 
   Rcpp::List list_out = xml_utils::encoding_decoding_test_function(input);
